clean headers
