+++ src/input_handler/syntax_list_utils.c
@@ -6,7 +6,7 @@
 /*   By: sebasnadu <johnavar@student.42berlin.de>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/15 17:32:07 by sebasnadu         #+#    #+#             */
-/*   Updated: 2023/11/15 17:53:05 by sebasnadu        ###   ########.fr       */
+/*   Updated: 2023/11/23 20:54:44 by sebasnadu        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
