+++ src/signals/signals.c
@@ -6,7 +6,7 @@
 /*   By: migmanu <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/06 15:22:11 by migmanu           #+#    #+#             */
-/*   Updated: 2023/11/22 16:01:02 by johnavar         ###   ########.fr       */
+/*   Updated: 2023/11/23 20:53:47 by sebasnadu        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,11 +14,11 @@
 #include <readline/readline.h>
 #include <signal.h>
 
-void	handle_sigint(int signum)
+static void	handle_sigint(int signum)
 {
 	if (signum == SIGINT)
 	{
-		g_exit_status = 130;
+		g_exit_status = 128 + signum;
 		ioctl(STDIN_FILENO, TIOCSTI, "\n");
 		rl_replace_line("", 0);
 		rl_on_new_line();
